feat: 뉴-프로젝트 카드 UI 모바일버전 추가
